requirements.txt
Pillow==9.5.0
numpy==1.24.3
pycryptodome==3.18.0
argparse==1.4.0
